Fix some chunking issues, extract CalendarEventsRepository
